Move export structs to pkg and document them
