+++ src/internal/data/data_collection.go
@@ -159,24 +159,3 @@ func StateOf(prev, curr path.Path) CollectionState {
 
 	return NotMovedState
 }
-
-// ExportCollection is the interface that is returned to the SDK consumer
-type ExportCollection interface {
-	// GetBasePath gets the base path of the collection
-	GetBasePath() string
-
-	// GetItems gets the items within the collection(folder)
-	GetItems(context.Context) <-chan ExportItem
-}
-
-type ExportItemData struct {
-	Name string
-	Body io.ReadCloser
-}
-
-// ExportItem is the item that is returned to the SDK consumer
-type ExportItem struct {
-	ID    string
-	Data  ExportItemData
-	Error error
-}
